expression_parser: fix precedence of logical 'not'
Logical 'not' should have higher precedence than logical 'and' but lower
precedence than comparison. For some reason it was implemented as having
equal precedence to unary plus/minus, which was not correct and caused
template code like the following to produce unexpected results:

```
{% if not foo is defined %}
    {% set foo = false %}
{% endif %}
```

Before, Jinja2Cpp would evaluate this as `if (not foo) is defined`,
which would always be true.

Signed-off-by: Jared Van Bortel <jared@nomic.ai>